Move getOwner to utils
